// src/app/shared/models/login.model.ts
export class LoginModel {
  UserId?: number;
  UserName?: string;
  Passwd?: string;
  AccessType?: string;
  RolliD?: number;
  InActive?: number;
  companyID?: number;
  TeamHeadID?: number;
  ModuleID?: number;
  ULvl?: number;
}

export class labelLoginModel {
  LabelId: number;
  ControlName: string;
  Describtion: string;
  controlCaption: string;
  controlName: string;
  ControlCaption: string;
}